frontend/requirements.txt
requests==2.32.4
streamlit==1.49.0
pillow
streamlit-ace==0.1.1
